From cypress, load new html page, invoke web worker, run tests, log output, assert. Add gha pipeline for web worker tests.
